[ADD] contest warning component, [UPD] form to split logic to report form
