+++ src/templates/ReportForm.js
@@ -25,6 +25,73 @@ const initialState = {
   ],
 };
 
+const emailField = {
+  name: "email",
+  label: "Email address",
+  helpText: "Used to send a copy of this form for your records",
+  widget: "text",
+  required: true,
+};
+
+const addressField = {
+  name: "polygonAddress",
+  label: "Polygon address",
+  helpText:
+    "Address where your prize should go. If you use a smart contract wallet, please contact one of our organizers in Discord in addition to adding the address here.",
+  widget: "text",
+  required: true,
+};
+
+const titleField = {
+  name: "title",
+  label: "Title",
+  helpText:
+    "Summarize your findings for the bug or vulnerability. (This will be the issue title.)",
+  widget: "text",
+  required: true,
+};
+
+const riskField = {
+  name: "risk",
+  label: "Risk rating",
+  widget: "select",
+  required: true,
+  options: [
+    {
+      label: "Gas Optimizations",
+      value: "G (Gas Optimization)",
+    },
+    {
+      label: "QA Report (low / non-critical)",
+      value: "QA (Quality Assurance)",
+    },
+    {
+      label: "Medium Risk",
+      value: "2 (Med Risk)",
+    },
+    {
+      label: "High Risk",
+      value: "3 (High Risk)",
+    },
+  ],
+};
+
+const vulnerabilityDetailsField = {
+  name: "details",
+  label: "Vulnerability details",
+  helpText: "Link to all referenced sections of code in GitHub",
+  widget: "textarea",
+  required: true,
+};
+
+const qaGasDetailsField = {
+  name: "qaGasDetails",
+  label: "Vulnerability details",
+  helpText: "Link to all referenced sections of code in GitHub",
+  widget: "textarea",
+  required: true,
+};
+
 const ReportForm = (props) => {
   const endTime = props.data.contestsCsv.end_time;
   const hasContestEnded = Date.now() > new Date(endTime).getTime();
@@ -48,6 +115,12 @@ const ReportForm = (props) => {
           contest={props.data.contestsCsv.contestid}
           sponsor={props.data.contestsCsv.sponsor.name}
           initialState={initialState}
+          emailField={emailField}
+          addressField={addressField}
+          titleField={titleField}
+          riskField={riskField}
+          vulnerabilityDetailsField={vulnerabilityDetailsField}
+          qaGasDetailsField={qaGasDetailsField}
         />
       )}
     </main>
